Retrieve calanguize_genomes script from the Github repository

This script downloads the *calanguize_genomes.pl* Perl script from the repository, together with associated README instructions for using the script, managing dependencies, etc. It will extract the data into a folder containing everything that is needed for preparing data for using CALANGO.

retrieve_calanguize_genomes(
  target.dir,
  method = "auto",
  unzip = getOption("unzip")
)

Arguments

target.dir
path to the folder where the files will be saved ( accepts relative and absolute paths)

method
Method to be used for downloading files. Current download methods are "internal", "wininet" (Windows only) "libcurl", "wget" and "curl", and there is a value "auto": see _Details_ and _Note_ in the documentation of utils::download.file().

unzip
The unzip method to be used. See the documentation of utils::unzip() for details.

Value

No return value, called for side effects (see Description).

Details

If the `target.dir` provided does not exist it is created (recursively) by the function.

Examples

if (FALSE) {
  CALANGO::retrieve_calanguize_genomes(target.dir = "./data")
}